pyproject.toml
[project]
name = "test-app"
version = "0.1.0"
description = "Unified file management across Dropbox, Box, Google Drive with Slack notifications, document parsing, and calendar integration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "dropbox>=12.0.0",
    "boxsdk>=3.12.0",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "slack-sdk>=3.33.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
